Initial commit
This is a Cucumber based sample test suite and framework utilizing the TestCentricity™ Web gem and a page-object model architecture for desktop and mobile web functional testing of the Appium Pro Chat web portal. This example is intended to demonstrate the instantiation of multiple WebDriver instances within a test case or test scenario to support the verification of real-time multiple user interactions across multiple concurrent browsers.